change main.rb and create bot.rb file
